genmesh: MULTIPLE CHANGES (see details)
Cleanup tests. For some reason, some randomly generated packings were
creating meshes which led to the weird error with 0 mrng values in some
cases. It's probably not related *only* to the packings, but this
somehow helps.

The other thing to check for the above problem is whether bead surfaces
include cut surfaces or not.

checkperiodic and genmprd had some small message modifications that got
mixed up by accident.